Small improvement to InterlockedExchange on ARM
Use movs instead of mov when setting the result in r0. This way the Z
flag will be set for the calling function which might allow some smaller
optimizations later on. It does not affect current code in any way,
because flags are not expected to be used across function calls.

git-svn-id: https://svn.freepascal.org/svn/fpc/trunk@22033 3ad0048d-3df7-0310-abae-a5850022a9f2